+++ src/const.ts
@@ -148,3 +148,9 @@ export enum NameSpace {
   Data = 'DATA',
   User = 'USER',
 }
+
+export enum PostReviewState {
+  Sending = 'Sending',
+  Send = 'Send',
+  Error = 'Error',
+}
